+++ ui/js/page/fileListDownloaded/index.js
@@ -5,6 +5,11 @@ import {
   selectFileInfosDownloaded,
   selectIsFetchingFileListDownloadedOrPublished,
 } from "selectors/file_info";
+import {
+  selectMyClaimsWithoutChannels,
+  selectIsFetchingClaimListMine,
+} from "selectors/claims";
+import { doFetchClaimListMine } from "actions/content";
 import { doNavigate } from "actions/app";
 import { doCancelAllResolvingUris } from "actions/content";
 import FileListDownloaded from "./view";
@@ -12,6 +17,8 @@ import FileListDownloaded from "./view";
 const select = state => ({
   fileInfos: selectFileInfosDownloaded(state),
   isFetching: selectIsFetchingFileListDownloadedOrPublished(state),
+  claims: selectMyClaimsWithoutChannels(state),
+  isFetchingClaims: selectIsFetchingClaimListMine(state),
 });
 
 const perform = dispatch => ({
@@ -19,6 +26,7 @@ const perform = dispatch => ({
   fetchFileInfosDownloaded: () =>
     dispatch(doFetchFileInfosAndPublishedClaims()),
   cancelResolvingUris: () => dispatch(doCancelAllResolvingUris()),
+  fetchClaims: () => dispatch(doFetchClaimListMine()),
 });
 
 export default connect(select, perform)(FileListDownloaded);